Dökümantasyon ve hata kontrolü sistemleri tamamlandı ApiEndpoints.md dosyasına product api'ı tamamen eklendi Görsel ekleme Endpoint'i tamamlandı Satın alma Endpoint'i tamamlandı Artırma yapma Endpoint'i tamamlandı Ürün oluşturulurken kuralsız yapıda detay ekleyebilme desteği getirildi

+++ Controllers/Api/Account.cs
@@ -48,7 +48,7 @@ public class Account : ControllerBase
 
         var token = new JwtToken()
         {
-            UserID = (int)user.UserId,
+            UserID = user.UserId,
             UserRole = (byte)user.UserRole,
             Expiration = DateTime.UtcNow.AddHours(1)
         };
